Reconfiguring @jayallen’s commit f8f6423699bf413aa3b56010f73604b6612995bf in which he added better error reporting to the sub `get_group_ids` used by the `*IfFieldGroup` template tags.
Previously, specifying in `*IfFieldGroup` tags a group that did not exist in the current blog would suppress the build error and instead log to the activity log. The `*FieldGroup` tags still produced the same opaque error message @jayallen’s commit was intended to solve.

Now, specifying in `*IfFieldGroup` tags a group that does not exist in the current blog will simply cause the tags to output nothing, just as the `IfCategory` tag does if the `category` attribute is invalid. Further, now the `*FieldGroup` tags both log to the activity log and throw a more helpful build error.

This way, the conditional tags can be used on global templates without needing to wrap them in logic to prevent firing on blogs without the proper field groups, and additionally will not clog the activity log with messages in those cases. The functional block tags are now the ones that throw a build error, aiding in template development.